Tutorial5_Serverless/terraform/src/requirements.txt
azure-functions
azure-cosmos
azure-ai-vision-imageanalysis
azure-core